added code for Authentication and Authorisation
added code for Authentication and Authorisation

+++ src/main/java/com/hitakshi/project/airBnbApp/service/RoomServiceImpl.java
@@ -3,12 +3,15 @@ package com.hitakshi.project.airBnbApp.service;
 import com.hitakshi.project.airBnbApp.dto.RoomDto;
 import com.hitakshi.project.airBnbApp.entity.Hotel;
 import com.hitakshi.project.airBnbApp.entity.Room;
+import com.hitakshi.project.airBnbApp.entity.User;
 import com.hitakshi.project.airBnbApp.exception.ResourceNotFoundException;
+import com.hitakshi.project.airBnbApp.exception.UnAuthorisedException;
 import com.hitakshi.project.airBnbApp.repository.HotelRepository;
 import com.hitakshi.project.airBnbApp.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -33,6 +36,10 @@ public class RoomServiceImpl implements RoomService{
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id : " +hotelId));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
         Room room = modelMapper.map(roomDto,Room.class);
         room.setHotel(hotel);
         room = roomRepository.save(room);
@@ -52,6 +59,11 @@ public class RoomServiceImpl implements RoomService{
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id : " +hotelId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
 
         return hotel.getRooms()
                 .stream()
@@ -79,6 +91,11 @@ public class RoomServiceImpl implements RoomService{
         if(!isExits) throw new ResourceNotFoundException("Room not found with id :"+id);*/
         Room room = roomRepository.findById(id)
                 .orElseThrow(()-> new ResourceNotFoundException("Room not found with id "));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner())) {
+            throw new UnAuthorisedException("This user does not own this room with id: "+id);
+        }
 
         inventoryService.deleteAllInventoriesOfRoom(room);
         roomRepository.deleteById(id);
